cleanup SSML tests to make similar to others
